move recovery lock to cmd_results file
